Don't await disconnection forever
There was a bug in PubSubClient::stop() - a surplus call to
`input_state()`. Also, it now waits for data shorter, not the usual
"full (default) time" . This time is not yet user settable.

Since ESP8266 can, in some weird and as-of-yet-unexplained situations,
take a _long_ time (hours, even) to change the state of the
`WiFiClient` to "not connected" after a `stop()`, we no longer await
this disconnection "forever".  There is now a timeout (for now, not
user-settable) and we give up after a short while. This might lead to
some (probably temporary) leaks, but, we haven't run into problems in
our testing.